Removed roster.
Want to know who goes to Hacks? Show up to Hacks.

Apparently someone name dropped Hacks during an interview to sound
cool, and said he liked that we had free pizza. The interviewer was a
friend of mine and knew that we don't do free food (*) and called him
out on it.

I'm really not sure what to make of this.

---

(*) Not necessarily true -- on Valentines Day I brought in 100 cookies I
baked the night before